+++ web-app/src/Components/Sights/New/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import SightForm from './Form';
-import ContainerLayout from '../UI/ContainerLayout';
+import ContainerLayout from 'Components/UI/ContainerLayout';
 
 const CreateSight = () => (
   <ContainerLayout header="Добавяне на обект">
